refactor: component naming and tests
