+++ module1/src/components/GenreSelector/genre-selector.module.scss
