+++ module1/src/components/index.tsx
@@ -1,0 +1,3 @@
+export {GenreSelector} from "./GenreSelector/genre-selector";
+export {SearchForm} from "./SearchForm/search-form";
+
