+++ module1/src/constants/сonstants.tsx
